requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
--extra-index-url https://repo.fury.io/rWCQLV1mdZaKwQuV-PFA/roverdotcom/

coverage==7.4.1
    # via pytest-cov
docutils==0.13.1
    # via -r requirements.in
flake8==7.0.0
    # via -r requirements.in
freezegun==1.1.0
    # via -r requirements.in
isort==5.13.2
    # via -r requirements.in
multilint==2.1.0
    # via -r requirements.in
py==1.4.34
    # via pytest
pygments==2.2.0
    # via -r requirements.in
pytest==8.0.0
    # via
    #   -r requirements.in
    #   pytest-cov
    #   pytest-django
pytest-cov==4.1.0
    # via -r requirements.in
pytest-django==4.8.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via freezegun
pytz==2021.3
    # via -r requirements.in
six==1.10.0
    # via python-dateutil
sqlparse==0.4.4
    # via -r requirements.in
black==24.1.1

# Nexus
https://github.com/roverdotcom/nexus/archive/3ec4808cf45b1c30e083f90ac1e33e9a3d0a7570.zip

# The following packages are considered to be unsafe in a requirements file:
# setuptools
